OY-3623 : Refactor utils
